+++ app/controllers/submissions.ctrl.js
@@ -70,10 +70,6 @@ const createSubmission = async (req, res, next) => {
     reCaptchaValue
   } = req.body;
 
-  if (!salesforce_id || salesforce_id === undefined) {
-    salesforce_id = res.locals.sf_contact_id;
-  }
-
   const requiredFields = [
     "submission_date",
     "birthdate",
@@ -96,57 +92,61 @@ const createSubmission = async (req, res, next) => {
 
   const missingField = requiredFields.find(field => !(field in req.body));
   if (!terms_agree) {
+    console.error(`submissions.ctrl.js > 99: !termsAgree`);
     return res.status(422).json({
       reason: "ValidationError",
       message: "Must agree to terms of service"
     });
   } else if (missingField) {
-    // console.log(`submissions.ctrl.js > 105: missing ${missingField}`);
+    console.error(`submissions.ctrl.js > 105: missing ${missingField}`);
     return res.status(422).json({
       reason: "ValidationError",
       message: `Missing required field ${missingField}`
     });
-  } else if (process.env.NODE_ENV !== "testing") {
-    verifyHumanity(reCaptchaValue, ip_address).catch(err => {
-      console.log(err);
-      return res
-        .status(422)
-        .json({ message: "Please verify that you are a human" });
-    });
   }
-  const createSubmissionResult = await submissions.createSubmission(
-    ip_address,
-    submission_date,
-    agency_number,
-    birthdate,
-    cell_phone,
-    employer_name,
-    first_name,
-    last_name,
-    home_street,
-    home_city,
-    home_state,
-    home_zip,
-    home_email,
-    preferred_language,
-    terms_agree,
-    signature,
-    text_auth_opt_out,
-    online_campaign_source,
-    legal_language,
-    maintenance_of_effort,
-    seiu503_cba_app_date,
-    direct_pay_auth,
-    direct_deposit_auth,
-    immediate_past_member_status,
-    salesforce_id
-  );
 
-  if (!createSubmissionResult || createSubmissionResult.message) {
-    // console.log(
-    //   `submissions.ctrl.js > 135: ${createSubmissionResult.message ||
-    //     "There was an error saving the submission"}`
-    // );
+  const createSubmissionResult = await submissions
+    .createSubmission(
+      ip_address,
+      submission_date,
+      agency_number,
+      birthdate,
+      cell_phone,
+      employer_name,
+      first_name,
+      last_name,
+      home_street,
+      home_city,
+      home_state,
+      home_zip,
+      home_email,
+      preferred_language,
+      terms_agree,
+      signature,
+      text_auth_opt_out,
+      online_campaign_source,
+      legal_language,
+      maintenance_of_effort,
+      seiu503_cba_app_date,
+      direct_pay_auth,
+      direct_deposit_auth,
+      immediate_past_member_status,
+      salesforce_id
+    )
+    .catch(err => {
+      console.error(`submissions.ctrl.js > 136: ${err.message}`);
+      return res.status(500).json({ message: err.message });
+    });
+
+  if (
+    !createSubmissionResult ||
+    !createSubmissionResult.length ||
+    createSubmissionResult.message
+  ) {
+    console.error(
+      `submissions.ctrl.js > 142: ${createSubmissionResult.message ||
+        "There was an error saving the submission"}`
+    );
     return res.status(500).json({
       message:
         createSubmissionResult.message ||
@@ -154,9 +154,12 @@ const createSubmission = async (req, res, next) => {
     });
   } else {
     res.locals.submission_id = createSubmissionResult[0].id;
-    return res
-      .status(200)
-      .json({ salesforce_id, submission_id: createSubmissionResult[0].id });
+    res.locals.currentSubmission = createSubmissionResult[0];
+    return res.status(200).json({
+      salesforce_id,
+      submission_id: createSubmissionResult[0].id,
+      currentSubmission: createSubmissionResult[0]
+    });
   }
 };
 
@@ -168,44 +171,48 @@ const createSubmission = async (req, res, next) => {
 const updateSubmission = async (req, res, next) => {
   const updates = req.body;
   const { id } = req.params;
-  // console.log(`subm.ctrl.js > 173: ${id}`);
+  // console.log(`subm.ctrl.js > 173 - id: ${id} (updates below)`);
   // console.log(updates);
-  try {
-    if (!updates || !Object.keys(updates).length) {
-      return res.status(422).json({ message: "No updates submitted" });
-    }
-    if (!id) {
-      return res.status(422).json({ message: "No Id Provided in URL" });
-    }
 
-    const updateSubmissionResult = await submissions.updateSubmission(
-      id,
-      updates
-    );
+  if (!updates || !Object.keys(updates).length) {
+    console.error("subm.ctrl.js > 173: !updates");
+    return res.status(422).json({ message: "No updates submitted" });
+  }
+  if (!id) {
+    console.error("subm.ctrl.js > 177: !id");
+    return res.status(422).json({ message: "No Id Provided in URL" });
+  }
 
-    if (
-      !updateSubmissionResult ||
-      updateSubmissionResult.message ||
-      updateSubmissionResult.length === 0
-    ) {
-      const errmsg =
-        updateSubmissionResult.message ||
-        "There was an error updating the submission";
-      // console.error(`submissions.ctrl.js > 205: ${errmsg}`);
+  const updateSubmissionResult = await submissions
+    .updateSubmission(id, updates)
+    .catch(err => {
+      console.error(`subm.ctrl.js > 188: ${err.message}`);
       return res.status(500).json({
-        message: errmsg
+        message: err.message
       });
-    } else {
-      // console.log("subm.ctrl.js > 201: returning to client");
-      // saving to res.locals to make id available for testing
-      res.locals.submission_id = updateSubmissionResult[0].id;
-      return res
-        .status(200)
-        .json({ submission_id: updateSubmissionResult[0].id });
-    }
-  } catch (error) {
-    // console.error(`submissions.ctrl.js > 217: ${error}`);
-    return res.status(404).json({ message: error.message });
+    });
+
+  if (
+    !updateSubmissionResult ||
+    (updateSubmissionResult && updateSubmissionResult.message) ||
+    updateSubmissionResult.length === 0
+  ) {
+    const errmsg =
+      updateSubmissionResult && updateSubmissionResult.message
+        ? updateSubmissionResult.message
+        : "There was an error updating the submission";
+    console.error(`submissions.ctrl.js > 205: ${errmsg}`);
+    return res.status(404).json({
+      message: errmsg
+    });
+  } else if (updateSubmissionResult[0] && updateSubmissionResult[0].id) {
+    // console.log("subm.ctrl.js > 201: returning to client");
+    // console.log(updateSubmissionResult[0].id);
+    // saving to res.locals to make id available for testing
+    res.locals.submission_id = updateSubmissionResult[0].id;
+    return res
+      .status(200)
+      .json({ submission_id: updateSubmissionResult[0].id });
   }
 };
 
@@ -215,15 +222,24 @@ const updateSubmission = async (req, res, next) => {
  */
 const deleteSubmission = async (req, res, next) => {
   let result;
+  const userType = req.user.type;
+  if (userType != "admin" || !userType) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to access this content. Please consult an administrator."
+    });
+  }
   try {
     result = await submissions.deleteSubmission(req.params.id);
     if (result.message === "Submission deleted successfully") {
       return res.status(200).json({ message: result.message });
     }
+    console.error(`submissions.ctrl.js > 225: ${result.message}`);
     return res.status(500).json({
       message: "An error occurred and the submission was not deleted."
     });
   } catch (err) {
+    console.error(`submissions.ctrl.js > 229: ${err.message}`);
     return res.status(500).json({ message: err.message });
   }
 };
@@ -232,6 +248,13 @@ const deleteSubmission = async (req, res, next) => {
  *  @returns  {Array|Object}   Array of submission objects OR error message
  */
 const getSubmissions = (req, res, next) => {
+  const userType = req.user.type;
+  if (!["admin", "view", "edit"].includes(userType)) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to access this content. Please consult an administrator."
+    });
+  }
   return submissions
     .getSubmissions()
     .then(submissions => {
@@ -239,7 +262,10 @@ const getSubmissions = (req, res, next) => {
       res.locals.testData = submissions[0];
       return res.status(200).json(submissions);
     })
-    .catch(err => res.status(500).json({ message: err.message }));
+    .catch(err => {
+      console.error(`submissions.ctrl.js > 247: ${err.message}`);
+      res.status(500).json({ message: err.message });
+    });
 };
 
 /** Get one submission
@@ -247,10 +273,18 @@ const getSubmissions = (req, res, next) => {
  *  @returns  {Object}        Submission object OR error message.
  */
 const getSubmissionById = (req, res, next) => {
+  const userType = req.user.type;
+  if (!["admin", "view", "edit"].includes(userType)) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to access this content. Please consult an administrator."
+    });
+  }
   return submissions
     .getSubmissionById(req.params.id)
     .then(submission => {
       if (!submission || submission.message) {
+        console.error(`submissions.ctrl.js > 261: ${submission.message}`);
         return res
           .status(404)
           .json({ message: submission.message || "Submission not found" });
@@ -260,7 +294,10 @@ const getSubmissionById = (req, res, next) => {
         return res.status(200).json(submission);
       }
     })
-    .catch(err => res.status(404).json({ message: err.message }));
+    .catch(err => {
+      console.error(`submissions.ctrl.js > 272: ${err.message}`);
+      res.status(500).json({ message: err.message });
+    });
 };
 
 /**
@@ -269,40 +306,36 @@ const getSubmissionById = (req, res, next) => {
  * @param {String} ip_address users ipAdress
  * @returns {Bool} returns true for human, false for bot
  */
-const verifyHumanity = (token, ip_address) => {
-  return new Promise((resolve, reject) => {
-    const key =
-      process.env.NODE_ENV === "testing"
-        ? process.env.TEST_RECAPTCHA_SECRET_KEY
-        : process.env.RECAPTCHA_SECRET_KEY;
-    return request.post(
-      "https://www.google.com/recaptcha/api/siteverify",
-      {
-        form: {
-          secret: key,
-          response: token,
-          remoteip: ip_address
-        }
-      },
-      (err, httpResponse, body) => {
-        if (err) {
-          // console.log(err);
-          reject(new Error(err));
+const verifyHumanity = async (req, res) => {
+  // console.log(`verifyHumanity`);
+  const { token, ip_address } = req.body;
+  const key = process.env.RECAPTCHA_V3_SECRET_KEY;
+  return request.post(
+    "https://www.google.com/recaptcha/api/siteverify",
+    {
+      form: {
+        secret: key,
+        response: token,
+        remoteip: ip_address
+      }
+    },
+    (err, httpResponse, body) => {
+      if (err) {
+        console.error(`submissions.ctrl.js > 298: ${err}`);
+        return res.status(500).json({ message: err.message });
+      } else {
+        const r = JSON.parse(body);
+        if (r.success) {
+          // console.log(`submissions.ctrl.js > 297: recaptcha score: ${r.score}`);
+          return res.status(200).json({ score: r.score });
         } else {
-          const r = JSON.parse(body);
-          // console.log(r['error-codes']);
-          if (r.success) {
-            // console.log(`submissions.ctrl.js > 291`);
-            // console.log(r.success);
-            resolve(r.success);
-          } else {
-            // console.log(`submissions.ctrl.js > 295`);
-            reject(new Error(`reCaptcha Error: ${r["error-codes"][0]}`));
-          }
+          console.error(`submissions.ctrl.js > 306: recaptcha failure`);
+          console.error(r["error-codes"][0]);
+          return res.status(500).json({ message: r["error-codes"][0] });
         }
       }
-    );
-  });
+    }
+  );
 };
 
 /* ================================ EXPORT ================================= */
