+++ app/controllers/users.ctrl.js
@@ -12,25 +12,52 @@ const users = require("../../db/models/users");
 /** Create a new user
  *  @param    {String}   name            Name from google profile.
  *  @param    {String}   email           Email from google profile.
+ *  @param    {String}   userType        User access type.
  *  @param    {String}   avatar_url      Picture from google profile.
  *  @param    {String}   google_id       Google unique ID.
  *  @param    {String}   google_token    Google auth token.
  *  @returns  {Object}                   New user object OR error message.
  */
 const createUser = (req, res, next) => {
-  const { name, email, avatar_url, google_id, google_token } = req.body;
-  if (name && email) {
+  const {
+    name,
+    email,
+    type,
+    avatar_url,
+    google_id,
+    google_token,
+    requestingUserType
+  } = req.body;
+  // console.log(req.body);
+  if (requestingUserType != "admin" || !requestingUserType) {
+    // console.log(`users.ctrl.js > 34`);
+    return res.status(500).json({
+      message:
+        "You do not have permission to do this. Please consult an administrator."
+    });
+  }
+  if (name && email && type) {
     return users
-      .createUser(name, email, avatar_url, google_id, google_token)
+      .createUser(name, email, avatar_url, google_id, google_token, type)
       .then(users => {
         const user = users[0];
-        res.status(200).json(user);
+        res.locals.testData = user;
+        return res.status(200).json(user);
       })
       .catch(err => {
-        console.log(`users.ctrl.js > 30: ${err}`);
-        res.status(500).json({ message: err.message });
+        console.log(`users.ctrl.js > 48: ${err.message}`);
+        let message = err.message;
+        if (
+          err.message.includes(
+            'duplicate key value violates unique constraint "users_email_unique"'
+          )
+        ) {
+          message = `A user with email ${email} already exists.`;
+        }
+        res.status(500).json({ message });
       });
   } else {
+    console.log(`users.ctrl.js > 61: missing required field`);
     return res
       .status(500)
       .json({ message: "There was an error creating the user account" });
@@ -42,29 +69,45 @@ const createUser = (req, res, next) => {
  *  @param    {Object}   updates         Key/value pairs for fields to update.
  ****  @param    {String}   name        Updated name.
  ****  @param    {String}   email       Updated email.
+ ****  @param    {String}   userType    Updated User access type.
  ****  @param    {String}   avatar_url  Updated avatar_url.
  *  @returns  {Object}                   Updated user object OR error message.
  */
 const updateUser = (req, res, next) => {
+  // console.log(req.body);
   const { updates } = req.body;
   const { id } = req.params;
-  if (!updates || !Object.keys(updates).length) {
-    return res.status(404).json({ message: "No updates submitted" });
+  const requestingUserType = req.user.type;
+  if (requestingUserType != "admin" || !requestingUserType) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to do this. Please consult an administrator."
+    });
   }
-
+  if (!id) {
+    return res.status(422).json({ message: "No Id Provided in URL" });
+  }
+  if (!updates || !Object.keys(updates).length) {
+    return res.status(422).json({ message: "No updates submitted" });
+  }
   return users
     .updateUser(id, updates)
     .then(user => {
+      // console.log(user);
       if (user.message || !user) {
         return res.status(404).json({
           message:
-            user.message || "An error occured while trying to update this user"
+            user.message || "An error occurred while trying to update this user"
         });
       } else {
+        res.locals.testData = user[0];
         return res.status(200).json(user);
       }
     })
-    .catch(err => res.status(500).json({ message: err.message }));
+    .catch(err => {
+      console.log(`users.ctrl.js > 107: ${err.message}`);
+      return res.status(500).json({ message: err.message });
+    });
 };
 
 /** Delete user
@@ -72,6 +115,13 @@ const updateUser = (req, res, next) => {
  *  @returns  Success or error message.
  */
 const deleteUser = (req, res, next) => {
+  const requestingUserType = req.user.type;
+  if (requestingUserType != "admin" || !requestingUserType) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to do this. Please consult an administrator."
+    });
+  }
   return users
     .deleteUser(req.params.id)
     .then(result => {
@@ -83,17 +133,38 @@ const deleteUser = (req, res, next) => {
         });
       }
     })
-    .catch(err => res.status(404).json({ message: err.message }));
+    .catch(err => {
+      console.log(`users.ctrl.js > 139: ${err.message}`);
+      return res.status(404).json({ message: err.message });
+    });
 };
 
 /** Get all users
  *  @returns  {Array|Object}   Array of user objects OR error message
  */
 const getUsers = (req, res, next) => {
+  const userType = req.user.type;
+  if (!userType || userType !== "admin") {
+    return res.status(500).json({
+      message:
+        "You do not have permission to do this. Please consult an administrator."
+    });
+  }
   return users
     .getUsers()
-    .then(users => res.status(200).json(users))
-    .catch(err => res.status(404).json({ message: err.message }));
+    .then(result => {
+      // console.log(result);
+      if (!result || result.message) {
+        return res.status(404).json({ message: result.message });
+      } else {
+        res.locals.testData = result;
+        return res.status(200).json(result);
+      }
+    })
+    .catch(err => {
+      console.log(`users.ctrl.js > 168: ${err.message}`);
+      return res.status(404).json({ message: err.message });
+    });
 };
 
 /** Get one user
@@ -101,21 +172,58 @@ const getUsers = (req, res, next) => {
  *  @returns  {Object}        User object OR error message.
  */
 const getUserById = (req, res, next) => {
+  const { id } = req.params;
+  if (!id) {
+    console.log(`users.ctrl.js > 176: no Id in params`);
+    return res.status(500).json({ message: "No user Id provided" });
+  }
   return users
-    .getUserById(req.params.id)
+    .getUserById(id)
     .then(user => {
       if (!user || user.message) {
-        console.log(user.message);
-        return res
-          .status(404)
-          .json({ message: user.message || "User not found" });
+        let message = "User not found";
+        if (user && user.message) {
+          message = user.message;
+        }
+        return res.status(404).json({ message });
       } else {
-        res.status(200).json(user);
+        res.locals.testData = user;
+        return res.status(200).json(user);
       }
     })
-    .catch(err => res.status(404).json({ message: err.message }));
+    .catch(err => {
+      console.log(`users.ctrl.js > 196: ${err.message}`);
+      return res.status(500).json({ message: err.message });
+    });
 };
 
+/** Get one user
+ *  @param    {String}   email   Email of the requested user.
+ *  @returns  {Object}        User object OR error message.
+ */
+const getUserByEmail = (req, res, next) => {
+  const requestingUserType = req.user.type;
+  if (requestingUserType != "admin" || !requestingUserType) {
+    return res.status(500).json({
+      message:
+        "You do not have permission to do this. Please consult an administrator."
+    });
+  }
+  return users
+    .getUserByEmail(req.params.email)
+    .then(user => {
+      if (!user) {
+        return res.status(404).json({ message: "User not found" });
+      } else {
+        res.locals.testData = user;
+        return res.status(200).json(user);
+      }
+    })
+    .catch(err => {
+      console.log(`users.ctrl.js > 217: ${err.message}`);
+      return res.status(500).json({ message: err.message });
+    });
+};
 /* ================================ EXPORT ================================= */
 
 module.exports = {
@@ -123,5 +231,6 @@ module.exports = {
   updateUser,
   deleteUser,
   getUserById,
+  getUserByEmail,
   getUsers
 };
